Finish button migration
