+++ packages/base/src/features/layers/symbology/components/color_stops/StopContainer.tsx
@@ -2,7 +2,7 @@ import { UUID } from '@lumino/coreutils';
 import React from 'react';
 
 import { IStopRow } from '@/src/features/layers/symbology/symbologyDialog';
-import { ButtonTw } from '@/src/shared/components/ButtonTw';
+import { Button } from '@/src/shared/components/Button';
 import StopRow from './StopRow';
 
 interface IStopContainerProps {
@@ -55,9 +55,9 @@ const StopContainer: React.FC<IStopContainerProps> = ({
         ))}
       </div>
       <div className="jp-gis-symbology-button-container">
-        <ButtonTw variant="outline" size="sm" onClick={addStopRow}>
+        <Button variant="outline" size="sm" onClick={addStopRow}>
           Add Stop
-        </ButtonTw>
+        </Button>
       </div>
     </>
   );
