+++ packages/base/style/base.css
@@ -22,7 +22,6 @@
 @import url('./shared/propertyRows.css');
 @import url('./labels.css');
 @import url('./openeo.css');
-@import url('./shared/button.css');
 @import url('./shared/errorBanner.css');
 @import url('./shared/errorTip.css');
 
@@ -385,14 +384,6 @@ body[data-notebook='specta'] #main.jp-ThemedContainer {
 
 .jgis-center-content {
   text-align: center;
-}
-
-.jgis-button.jgis-bg-transparent {
-  background-color: transparent;
-}
-
-.jgis-button.jgis-bg-transparent:hover {
-  background-color: transparent;
 }
 
 .jgis-rotate-90 {
