--- packages/base/style/shared/button.css
@@ -1,172 +1,0 @@
-.jgis-button {
-  box-sizing: border-box;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  white-space: nowrap;
-  border-radius: var(--jp-border-radius, 0.375rem);
-  font-size: 0.875rem;
-  font-weight: 600;
-  outline: none;
-  background-color: var(--jp-layout-color0);
-  color: var(--jp-ui-font-color0);
-  border: 0.0375rem solid
-    color-mix(in srgb, var(--jp-border-color1), transparent 20%);
-  transition: background-color 0.2s ease;
-  height: 1.5rem;
-  padding: 1rem 0.5rem;
-  max-width: 20rem;
-  gap: 0.25rem;
-}
-
-.jgis-button:hover {
-  cursor: pointer;
-  background-color: color-mix(
-    in srgb,
-    var(--jp-layout-color2),
-    transparent 20%
-  );
-  transition: background-color 0.2s ease;
-}
-
-.jgis-button:focus-visible {
-  outline: none;
-  box-shadow: 0 0 0 3px
-    color-mix(in srgb, var(--jp-ui-font-color0), transparent 50%);
-}
-
-.jgis-button:disabled {
-  pointer-events: none;
-  opacity: 0.5;
-}
-
-.jgis-button:disabled:focus-visible {
-  outline: none;
-  box-shadow: none;
-}
-
-.jgis-button[data-variant='destructive'] {
-  background-color: var(--jp-error-color0);
-  color: var(--jp-ui-inverse-font-color0);
-  transition: background-color 0.2s ease;
-}
-
-.jgis-button[data-variant='destructive']:hover {
-  background-color: color-mix(in srgb, var(--jp-error-color0), transparent 10%);
-}
-
-.jgis-button[data-variant='outline'] {
-  border: 1px solid var(--jp-border-color0);
-  background-color: var(--jp-layout-color0);
-  color: var(--jp-ui-font-color0);
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease;
-}
-
-.jgis-button[data-variant='outline']:hover {
-  cursor: pointer;
-  background-color: color-mix(
-    in srgb,
-    var(--jp-ui-font-color0),
-    transparent 95%
-  );
-  color: var(--jp-ui-font-color2);
-}
-
-.jgis-button[data-variant='secondary'] {
-  background-color: var(--jp-layout-color2);
-  color: var(--jp-ui-font-color0);
-  transition: background-color 0.2s ease;
-}
-
-.jgis-button[data-variant='secondary']:hover {
-  background-color: color-mix(
-    in srgb,
-    var(--jp-layout-color2),
-    transparent 10%
-  );
-}
-
-.jgis-button[data-variant='ghost'] {
-  background-color: transparent;
-  border: 1px solid transparent;
-  color: var(--jp-ui-font-color0);
-  display: flex;
-  align-items: center;
-}
-
-.jgis-button[data-variant='ghost']:hover {
-  background-color: color-mix(
-    in srgb,
-    var(--jp-ui-font-color0),
-    transparent 90%
-  );
-}
-
-.jgis-button[data-variant='link'] {
-  color: var(--jp-ui-font-color0);
-  background-color: transparent;
-  text-decoration: none;
-  transition: text-decoration 0.2s ease;
-  border: none;
-  padding: 0;
-  height: max-content;
-  width: max-content;
-}
-
-.jgis-button[data-variant='link']:hover {
-  color: var(--jp-layout-color0);
-  text-decoration-color: var(--jp-ui-font-color0);
-}
-
-.jgis-button[data-variant='icon'] {
-  border: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 2rem;
-  width: 2rem;
-  padding: 0;
-  border-radius: 9999px;
-}
-
-/* sizing */
-.jgis-button[data-size='xs'] {
-  padding: 0.5rem 0.25rem;
-}
-
-.jgis-button[data-size='sm'] {
-  border-radius: var(--jp-border-radius, 0.375rem);
-  padding-left: 0.75rem;
-  padding-right: 0.75rem;
-}
-
-.jgis-button[data-size='lg'] {
-  height: 2.75rem;
-  border-radius: var(--jp-border-radius, 0.375rem);
-  padding-left: 2rem;
-  padding-right: 2rem;
-}
-
-.jgis-button[data-size='icon'] {
-  height: 2.5rem;
-  width: 2.5rem;
-}
-
-.jgis-button[data-size='icon-sm'] {
-  height: 1rem;
-  width: 1rem;
-}
-
-.jgis-button[data-size='icon-md'] {
-  height: 1.5rem;
-  width: 1.5rem;
-}
-
-.jgis-button[data-variant='icon'] > svg,
-.jgis-button[data-variant='icon'] > [data-icon] {
-  height: 1rem;
-  width: 1rem;
-  flex-shrink: 0;
-}
